Add files via upload
This report provides an analysis of renewable energy across various regions, utilizing Github as the platform for storing and sharing the information and also using Microsoft Excel for data cleaning, data analysis and also for the visualization. The report includes an overview of the renewable energy data, including total energy production and consumption, as well as a breakdown of renewable energy sources by region across different years. Additionally, the report includes visualizations of the data, such as maps and charts, to aid in understanding and interpreting the information. The report also shows the number of jobs opportunities created by the various energy sources and also the top countries for investment. Overall, this report aims to provide a comprehensive understanding of the renewable energy performance across different regions and help guide decision making for government, business and individuals.